Bumped version number to 0.03. Updated phpbb_login() to work with phpbb 3.0.*.
Previous implementation of phpbb_login() didn't work with phpbb 3.0.*.
* Login/logout actions are now handled by ucp.php?mode=login/logout not login.php and logout.php.
* It seems that at the time 0.02 was written, the login page contained only the login form.
  Now on every phpbb page there is an additional search form located in the upper right corner of the page
  submit_form tried to login using that search form instead of the login form. Adding form_id to
  submit_form fixed this issue.

There is an overall problem with phpBB.pm library. Most actions fail silently and without debug
information from log4perl it's not easy to see if the user logged in successfully or that any
plugin function call was executed correctly. I may be wrong with this judgment but just noting
my current concerns.